@@ -40,7 +40,7 @@ class Modus implements ArrayInterface
      *
      * @return array
      */
-    public function toOptionArray()
+    public function toOptionArray(): array
     {
         // @codingStandardsIgnoreStart
         $options = [
