@@ -49,15 +49,19 @@ class ActionUrlTest extends TestCase
         $expectsPostcodeNLUrl = 'http://test.nl/postcode/address/service/';
         $expectsPostcodeBEUrl = 'http://test.nl/postcode/address/service/be/getpostcode';
         $expectsStreetBEUrl   = 'http://test.nl/postcode/address/service/be/getstreet';
+        $expectsPostcodeDEUrl = 'http://test.nl/postcode/address/service/de/getpostcode';
+        $expectsStreetDEUrl   = 'http://test.nl/postcode/address/service/de/getstreet';
 
         $urlInterfaceMock = $this->getFakeMock(UrlInterface::class)->getMock();
         $expects          = $urlInterfaceMock->expects($this->exactly(3));
         $expects->method('getUrl')->withConsecutive(
             ['postcode/address/service', ['_secure' => true]],
             ['postcode/address/service/be/getpostcode', ['_secure' => true]],
-            ['postcode/address/service/be/getstreet', ['_secure' => true]]
+            ['postcode/address/service/be/getstreet', ['_secure' => true]],
+            ['postcode/address/service/de/getpostcode', ['_secure' => true]],
+            ['postcode/address/service/de/getstreet', ['_secure' => true]]
         );
-        $expects->willReturnOnConsecutiveCalls($expectsPostcodeNLUrl, $expectsPostcodeBEUrl, $expectsStreetBEUrl);
+        $expects->willReturnOnConsecutiveCalls($expectsPostcodeNLUrl, $expectsPostcodeBEUrl, $expectsStreetBEUrl, $expectsPostcodeDEUrl, $expectsStreetDEUrl);
 
         $instance = $this->getInstance(
             [
@@ -68,7 +72,9 @@ class ActionUrlTest extends TestCase
         $returns = [
             'postcode_service'        => $expectsPostcodeNLUrl,
             'postcode_be_getpostcode' => $expectsPostcodeBEUrl,
-            'postcode_be_getstreet'   => $expectsStreetBEUrl
+            'postcode_be_getstreet'   => $expectsStreetBEUrl,
+            'postcode_de_getpostcode' => $expectsPostcodeDEUrl,
+            'postcode_de_getstreet'   => $expectsStreetDEUrl
         ];
 
         $this->assertEquals($returns, $instance->getValue());
