@@ -42,16 +42,17 @@ class ApiConfigurationTest extends AbstractConfigurationTest
     private $base = 'https://api.postcodeservice.com/nl/';
     /** @var string  */
     private $beBase = 'https://api.postcodeservice.com/be/';
+    /** @var string  */
+    private $deBase = 'https://api.postcodeservice.com/de/';
+    /** @var string  */
+    private $version         = 'v5';
+    /** @var string  */
+    private $postcodeVersion = 'v3';
+    /** @var string  */
+    private $streetVersion   = 'v3';
 
     /** @var string  */
-    private $version         = 'v3';
-    /** @var string  */
-    private $postcodeVersion = 'v2';
-    /** @var string  */
-    private $streetVersion   = 'v2';
-
-    /** @var string  */
-    private $postcodeEndpoint = 'postcode-find/';
+    private $postcodeEndpoint = 'zipcode-find/';
     /** @var string  */
     private $streetEndpoint   = 'street-find/';
 
@@ -88,6 +89,15 @@ class ApiConfigurationTest extends AbstractConfigurationTest
     {
         $this->setXpath(ApiConfiguration::XPATH_API_BE_BASE, $this->beBase);
         $this->assertEquals($this->beBase, $this->instance->getBase('BE'));
+    }
+
+    /**
+     * @return void
+     */
+    public function testDeGetBase()
+    {
+        $this->setXpath(ApiConfiguration::XPATH_API_DE_BASE, $this->deBase);
+        $this->assertEquals($this->deBase, $this->instance->getBase('DE'));
     }
 
     /**
@@ -149,7 +159,29 @@ class ApiConfigurationTest extends AbstractConfigurationTest
         );
 
         $expected = $this->beBase . '/' . $this->postcodeVersion . '/';
-        $this->assertEquals($expected, $this->instance->getBeBaseUri($this->postcodeEndpoint));
+        $this->assertEquals($expected, $this->instance->getBEBaseUri($this->postcodeEndpoint));
+    }
+
+    /**
+     * @return void
+     */
+    public function testGetDeBasePostcodeUri()
+    {
+        $this->setXpathConsecutive(
+            [
+                ApiConfiguration::XPATH_API_DE_BASE,
+                ApiConfiguration::XPATH_API_DE_POSTCODE_VERSION,
+                ApiConfiguration::XPATH_API_DE_STREET_VERSION,
+            ],
+            [
+                $this->deBase,
+                $this->postcodeVersion,
+                $this->streetVersion
+            ]
+        );
+
+        $expected = $this->deBase . '/' . $this->postcodeVersion . '/';
+        $this->assertEquals($expected, $this->instance->getDEBaseUri($this->postcodeEndpoint));
     }
 
     /**
@@ -171,6 +203,28 @@ class ApiConfigurationTest extends AbstractConfigurationTest
         );
 
         $expected = $this->beBase . '/' . $this->streetVersion . '/';
-        $this->assertEquals($expected, $this->instance->getBeBaseUri($this->streetEndpoint));
+        $this->assertEquals($expected, $this->instance->getBEBaseUri($this->streetEndpoint));
+    }
+
+    /**
+     * @return void
+     */
+    public function testGetDeBaseStreetUri()
+    {
+        $this->setXpathConsecutive(
+            [
+                ApiConfiguration::XPATH_API_DE_BASE,
+                ApiConfiguration::XPATH_API_DE_POSTCODE_VERSION,
+                ApiConfiguration::XPATH_API_DE_STREET_VERSION,
+            ],
+            [
+                $this->deBase,
+                $this->postcodeVersion,
+                $this->streetVersion
+            ]
+        );
+
+        $expected = $this->deBase . '/' . $this->streetVersion . '/';
+        $this->assertEquals($expected, $this->instance->getDEBaseUri($this->streetEndpoint));
     }
 }
