@@ -1,4 +1,5 @@
 <?php
+
 /**
  *
  *          ..::..
@@ -18,13 +19,13 @@
  * It is available through the world-wide-web at this URL:
  * http://creativecommons.org/licenses/by-nc-nd/3.0/nl/deed.en_US
  * If you are unable to obtain it through the world-wide-web, please send an email
- * to support@postcodeservice.com so we can send you a copy immediately.
+ * to servicedesk@totalinternetgroup.nl so we can send you a copy immediately.
  *
  * DISCLAIMER
  *
  * Do not edit or add to this file if you wish to upgrade this module to newer
  * versions in the future. If you wish to customize this module for your
- * needs please contact support@postcodeservice.com for more information.
+ * needs please contact servicedesk@totalinternetgroup.nl for more information.
  *
  * @copyright   Copyright (c) Total Internet Group B.V. https://tig.nl/copyright
  * @license     http://creativecommons.org/licenses/by-nc-nd/3.0/nl/deed.en_US
@@ -33,10 +34,10 @@
 namespace TIG\Postcode\Webservices\Endpoints;
 
 // @codingStandardsIgnoreFile
-class GetAddress extends EndpointAbstract
+class GetDeStreet extends EndpointAbstract
 {
-    protected $endpoint     = 'find/';
-    protected $country      = 'NL';
-    protected $requestKeys  = ['zipcode', 'houseno']; // NL V5
-    protected $responseKeys = ['street', 'city']; // NL V5
+    protected $endpoint     = 'street-find/';
+    protected $country      = 'DE';
+    protected $requestKeys  = ['zipcode', 'city', 'street']; // DE V1
+    protected $responseKeys = ['street']; // DE V1
 }
