@@ -7,7 +7,7 @@ var config = {
             'Magento_Checkout/js/action/create-shipping-address': {
                 'TIG_Postcode/js/action/create-shipping-address-mixin': true
             },
-            'Magento_Checkout/js/action/set-billing-address' : {
+            'Magento_Checkout/js/action/set-billing-address': {
                 'TIG_Postcode/js/action/set-billing-address-mixin': true
             },
             'Magento_Checkout/js/action/place-order': {
@@ -19,17 +19,17 @@ var config = {
             'Magento_Checkout/js/action/set-payment-information': {
                 'TIG_Postcode/js/action/set-billing-address-mixin': true
             },
-
-            // Mixin for hiding TIG Attributes
-            'Magento_Checkout/js/view/billing-address' : {
+            
+            // Mixin for hiding Postcode Service TIG Attributes
+            'Magento_Checkout/js/view/billing-address': {
                 'TIG_Postcode/js/view/hide-tig-attributes-mixin': true
             },
-            'Magento_Checkout/js/view/shipping-address/address-renderer/default' : {
+            'Magento_Checkout/js/view/shipping-address/address-renderer/default': {
                 'TIG_Postcode/js/view/hide-tig-attributes-mixin': true
             },
-            'Magento_Checkout/js/view/shipping-information/address-renderer/default' : {
+            'Magento_Checkout/js/view/shipping-information/address-renderer/default': {
                 'TIG_Postcode/js/view/hide-tig-attributes-mixin': true
-            },
+            }
         }
     }
 };
