@@ -1,4 +1,35 @@
+/**
+ *
+ *          ..::..
+ *     ..::::::::::::..
+ *   ::'''''':''::'''''::
+ *   ::..  ..:  :  ....::
+ *   ::::  :::  :  :   ::
+ *   ::::  :::  :  ''' ::
+ *   ::::..:::..::.....::
+ *     ''::::::::::::''
+ *          ''::''
+ *
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Creative Commons License.
+ * It is available through the world-wide-web at this URL:
+ * http://creativecommons.org/licenses/by-nc-nd/3.0/nl/deed.en_US
+ * If you are unable to obtain it through the world-wide-web, please send an email
+ * to servicedesk@totalinternetgroup.nl so we can send you a copy immediately.
+ *
+ * DISCLAIMER
+ *
+ * Do not edit or add to this file if you wish to upgrade this module to newer
+ * versions in the future. If you wish to customize this module for your
+ * needs please contact servicedesk@totalinternetgroup.nl for more information.
+ *
+ * @copyright   Copyright (c) Total Internet Group B.V. https://tig.nl/copyright
+ * @license     http://creativecommons.org/licenses/by-nc-nd/3.0/nl/deed.en_US
+ */
 define([
+    // Importing required modules and dependencies
     'jquery',
     'mage/utils/wrapper',
     'Magento_Checkout/js/model/quote'
@@ -8,7 +39,7 @@ define([
     return function (setShippingInformationAction) {
         return wrapper.wrap(setShippingInformationAction, function (originalAction, messageContainer) {
 
-            if (messageContainer.custom_attributes != undefined) {
+            if (messageContainer.custom_attributes !== undefined) {
                 $.each(messageContainer.custom_attributes , function( key, value ) {
                     messageContainer['custom_attributes'][key] = value;
                 });
